requirements.txt
alt-profanity-check==0.24.0
asgiref==3.3.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
dj-database-url==0.5.0
django-cors-headers==3.7.0 
djangorestframework==3.12.4
Django==3.2.6
gunicorn==20.0.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
joblib==1.0.1
numpy==1.19.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
quotes==0.0.3
scikit-learn==0.24.0
scipy==1.5.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
whitenoise==5.2.0
zope.interface==5.2.0
